Extract Android startup recovery helpers
